+++ src/components/gantt-chart/_parts/bar-date-handle/BarDateHandle.module.css
