+++ src/components/gantt-chart/_parts/bar-progress-handle/BarProgressHandle.module.css.d.ts
