+++ src/components/gantt-chart/_parts/bar-progress-handle/type.ts
