+++ src/components/gantt-chart/_parts/bar/Bar.module.css
