+++ src/components/gantt-chart/_parts/bar/type.ts
@@ -1,4 +1,4 @@
-import type { Task } from "../../../../../../types/task";
+import type { Task } from "../../../../types/task";
 
 /**
  * @description Bar props
